third_party/blink/web_tests/platform/generic/fast/dom/Range/range-constructor-expected.txt

Tests that we can create instances of Range using new

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof new Range is "object"
PASS Object.prototype.toString.call(new Range) is "[object Range]"
PASS new Range instanceof Range is true
PASS Object.getPrototypeOf(new Range) is Range.prototype
PASS r.toString() is "Test"
PASS new innerWindow.Range instanceof innerWindow.Range is true
PASS Object.getPrototypeOf(new innerWindow.Range) is innerWindow.Range.prototype
PASS r.toString() is "Inner"
PASS successfullyParsed is true

TEST COMPLETE

Test
